requirements.txt
langchain-core
langgraph
langchain-groq
langchain-huggingface
langchain-pinecone
pinecone-client
langchain-community
wikipedia
arxiv
fastapi
pydantic
python-dotenv